requirements.txt
langchain
openai
python-dotenv
chromadb
beautifulsoup4
jsonlines
gradio
tiktoken